# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi REST API Wilayah Indonesia",
        "description": "REST API untuk menarik data provinsi, kota dan kabupaten, dan kecamatan yang ada di Indonesia dan juga dilengkapi dengan fitur pembuatan daftar nama konsumen via API",
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User Login",
                "description": "Generate API Token",
                "operationId": "Login",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/getprovinsi": {
            "get": {
                "tags": [
                    "Provinsi"
                ],
                "summary": "Mendapatkan Daftar Provinsi",
                "description": "Menampilkan semua daftar provinsi",
                "operationId": "getListProvinsi",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/getkota/{provinsi_id}": {
            "get": {
                "tags": [
                    "Kota"
                ],
                "summary": "Mendapatkan daftar Kota dan Kabupaten",
                "description": "Menampilkan daftar kota dan kabupaten di suatu provinsi",
                "operationId": "getListKota",
                "parameters": [
                    {
                        "name": "provinsi_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/getkecamatan/{id}": {
            "get": {
                "tags": [
                    "Kecamatan"
                ],
                "summary": "Mendapatkan daftar Kecamatan",
                "description": "Menampilkan daftar kecamatan di suatu kota atau kabupaten",
                "operationId": "getListKecamatan",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://api-zone.test//oauth/authorize",
                        "tokenUrl": "http://api-zone.test//oauth/token",
                        "refreshUrl": "http://api-zone.test//token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}